{{cookiecutter.project_slug}}/requirements/dreamhost.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt

# Python-MySQL Database Adapter
# ------------------------------------------------
mysqlclient==1.3.7

# Static and Media Storage
# ------------------------------------------------
django-storages-redux==1.3.2
{% if cookiecutter.use_whitenoise != 'y' -%}
Collectfast==0.2.3
{%- endif %}

# Mailgun Support
# ---------------
django-mailgun==0.9.1

{% if cookiecutter.use_sentry == "y" -%}
# Raven is the Sentry client
# --------------------------
raven
{%- endif %}

{% if cookiecutter.use_newrelic == "y" -%}
# Newrelic agent for performance monitoring
# -----------------------------------------
newrelic
{%- endif %}

{% if cookiecutter.use_opbeat == "y" -%}
# Opbeat agent for performance monitoring
# -----------------------------------------
opbeat
{%- endif %}
